Adds support for Nice Nano, no BLE yet.
