--- ffkb.md
@@ -1,1 +1,0 @@
-This is the firmware for the ffkb running on circuitpy
